Move split sub-package into vak/prep/ -- it's only used there
