+++ src/vak/prep/split/split.py
@@ -6,10 +6,11 @@ import logging
 import pathlib
 
 import numpy as np
+import pandas as pd
 
 from .algorithms import brute_force
 from .algorithms.validate import validate_split_durations
-from ..common.labels import from_df as labels_from_df
+from ...common.labels import from_df as labels_from_df
 
 
 logger = logging.getLogger(__name__)
